requirements.txt
websocket-client>=1.6.4
requests>=2.31.0
websockets>=12.0